SERVER/backend/requirements.txt
kafka-python==2.0.2
h5py==3.9.0
Werkzeug==2.2.2
flask==2.2.2
flask-cors==4.0.0